Add patch placement algorithm, flip functionality
Added a smarter patch placement algorithm. Currently just a greedy
algorithm, tries to maximize the number of open tiles with open adjacent
tiles. This encourages placing patches next to edges or patches that
have already been placed.

Also added the ability to flip a patch, in addition to rotating it.