Tweak imports, update base AI, add sampling run
Updated imports to not mess with sys path so everything can be imported
based on absolute path.

Updated basic AI to follow simple hand-crafted logic (buy patches with
econ gen early in the game, patches that fill space later in the game).
Right now this AI beats a random AI ~65% of the time. Looking to improve
on this, as this will be the benchmark for the learned AI.

Added a sampling run for AI v AI, so an integer can be given to the AI v
AI run, that number of games will be simulated, and P1/P2 win count will
be returned.