Add first cut of learned AI
Basic learned AI first draft. Right now it's just learning a general
value of each patch, so it has no concept of when to buy certain patches
(i.e, early mid or late game), how to deny patches, how to buy patches
that fit well on the board, how to balance time well, etc. etc.

Need to do some cleanup for managing what AI is running and things like
that. Should be able to export feature weights and load them in as well.